requirements.txt
absl-py==2.3.1
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.4
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
async-timeout==5.0.1
attrs==26.1.0
bitsandbytes==0.48.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
datasets==4.8.4
decord==0.6.0
diffusers @ git+https://github.com/huggingface/diffusers.git@6f1042e36cd588a7b66498f45c3bb7085e4fa395
dill==0.4.1
exceptiongroup==1.3.1
filelock==3.19.1
frozenlist==1.8.0
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
Jinja2==3.1.6
Markdown==3.10
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.19
networkx==3.3
numpy==2.1.2
nvidia-cublas==13.0.0.19
nvidia-cuda-cupti==13.0.48
nvidia-cuda-nvrtc==13.0.48
nvidia-cuda-runtime==13.0.48
nvidia-cudnn-cu13==9.13.0.50
nvidia-cufft==12.0.0.15
nvidia-cufile==1.15.0.42
nvidia-curand==10.4.0.35
nvidia-cusolver==12.0.3.29
nvidia-cusparse==12.6.2.49
nvidia-cusparselt-cu13==0.8.0
nvidia-ml-py @ file:///home/conda/feedstock_root/build_artifacts/nvidia-ml-py_1757635263657/work
nvidia-nccl-cu13==2.27.7
nvidia-nvjitlink==13.0.39
nvidia-nvshmem-cu13==3.3.24
nvidia-nvtx==13.0.39
nvitop @ file:///home/conda/feedstock_root/build_artifacts/nvitop_1765168836809/work
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.3
peft==0.18.0
pillow==11.3.0
platformdirs==4.5.0
propcache==0.4.1
protobuf==6.33.1
psutil==7.1.3
pyarrow==23.0.1
pydantic==2.12.4
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
pytz==2026.1.post1
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
sentencepiece==0.2.1
sentry-sdk==2.45.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tokenizers==0.22.1
torch==2.9.1+cu130
torchvision==0.24.1+cu130
tqdm==4.67.1
transformers==4.57.1
triton==3.5.1
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.5.0
wandb==0.23.0
Werkzeug==3.1.3
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.0
